refactor: load Alpine.js lazily
